Move the abstract diagnostic types to their own module
