+++ Sources/SwiftDiagnostics/DiagnosticMessage.swift
@@ -34,18 +34,3 @@ public protocol DiagnosticMessage {
   /// See ``DiagnosticMessageID``.
   var diagnosticID: DiagnosticMessageID { get }
 }
-
-/// A diagnostic how's ID is determined by the diagnostic's type.
-public protocol TypedDiagnosticMessage: DiagnosticMessage {
-  var diagnosticID: DiagnosticMessageID { get }
-}
-
-public extension TypedDiagnosticMessage {
-  static var diagnosticID: DiagnosticMessageID {
-    return DiagnosticMessageID("\(self)")
-  }
-
-  var diagnosticID: DiagnosticMessageID {
-    return Self.diagnosticID
-  }
-}
