+++ Sources/SwiftParser/Diagnostics/ParseDiagnosticsGenerator.swift
@@ -10,6 +10,7 @@
 //
 //===----------------------------------------------------------------------===//
 
+import SwiftDiagnostics
 import SwiftSyntax
 
 extension UnexpectedNodesSyntax {
