+++ Sources/swift-parser-test/swift-parser-test.swift
@@ -11,6 +11,7 @@
 //===----------------------------------------------------------------------===//
 
 
+import SwiftDiagnostics
 import SwiftSyntax
 import SwiftParser
 import Foundation
