+++ Tests/SwiftParserTest/DiagnosticAssertions.swift
@@ -1,3 +1,4 @@
+import SwiftDiagnostics
 import SwiftSyntax
 import SwiftParser
 import XCTest
